% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compRes}
\alias{compRes}
\title{Celsius conversion}
\usage{
compRes(x)
}
\arguments{
\item{X}{The expression matrix}
}
\value{
The compressed matrix
}
\description{
Compress a given gene expression matrix
}
\examples{
temp1 <- compRes(22);
}
