% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{domextr}
\alias{domextr}
\title{Celsius conversion}
\usage{
domextr(x, y)
}
\arguments{
\item{X}{The expression matrix}
}
\value{
Sigmoid matrix
 @examples 
temp1 <- compRes(22);
}
\description{
Find the cell types from any expression matrix
}
