% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{impRes}
\alias{impRes}
\title{A test package}
\usage{
impRes(x)
}
\arguments{
\item{Exp_mar}{The temperature in degrees Fahrenheit}
}
\value{
The expression matrix after imputation
}
\description{
Fahrenheit conversion
}
\details{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- normRes(50);
temp2 <- compRes(70) );
temp3 <- impRes(65)

}
