% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normRes}
\alias{normRes}
\title{Celsius conversion}
\usage{
normRes(x)
}
\arguments{
\item{X}{The expression matrix}
}
\value{
The compressed matrix
}
\description{
Find the cell types from any expression matrix
}
\examples{
temp1 <- compRes(22);
}
